+++ src/config.rs
@@ -1,21 +1,21 @@
 use std::{fs, io};
 
 use anyhow::{Context, Result, anyhow};
+use byte_unit::Byte;
 use camino::Utf8PathBuf;
 use serde_derive::Deserialize;
 
 use crate::backend::cache;
 
-// lol: Serde wants a function to call for defaults.
 #[inline]
-fn defcachesize() -> u64 {
+fn defcachesize() -> Byte {
     cache::DEFAULT_SIZE
 }
 
 #[derive(Debug, Deserialize)]
 pub struct Configuration {
     #[serde(default = "defcachesize")]
-    pub cache_size: u64,
+    pub cache_size: Byte,
 }
 
 impl Default for Configuration {
